refactor: rename agents/schemas to agents/models
The repo already has a top-level schemas/ directory holding the canonical
JSON Schema definitions for scenes and themes. Calling the agent-side
Pydantic types "schemas" too made it ambiguous which one was meant.
Rename to agents/models — these are Pydantic models used as result_type
on agents, not JSON Schema documents.
