+++ agents/models/__init__.py
@@ -1,7 +1,11 @@
-"""Pydantic schemas used as `output_type` by agents.
+"""Pydantic models used as `output_type` by agents.
 
-One module per schema family. Future lesson schemas live alongside
+One module per model family. Future lesson models live alongside
 `semantic_graph` here.
+
+Named ``models`` (not ``schemas``) to avoid collision with the repo-level
+``schemas/`` directory, which holds the canonical JSON Schema definitions
+for scenes and themes.
 """
 
 from .semantic_graph import (
