+++ agents/semantic_graph_enricher.py
@@ -11,7 +11,7 @@ import json
 from typing import Any, Dict, Optional
 
 from .base import BaseAgent
-from .schemas import SemanticGraph
+from .models import SemanticGraph
 
 
 _SYSTEM_PROMPT = """\
