+++ tests/agents/test_semantic_graph_enricher.py
@@ -12,7 +12,7 @@ from __future__ import annotations
 import pytest
 from pydantic import ValidationError
 
-from agents.schemas import SemanticGraph
+from agents.models import SemanticGraph
 from agents.semantic_graph_enricher import SemanticGraphEnrichmentAgent
